HW - 8 : Контекст + localStorage
