+++ src/components/CharacterCard/CharacterCard.js
@@ -20,7 +20,6 @@ const CharacterCard = ({
 }) => {
   const handleClick = () => {
     onLikeClick && onLikeClick(id);
-    console.log(id);
   };
 
   return (
